+++ src/Controller/ProductController.php
@@ -21,6 +21,11 @@ class ProductController
 
     public function getAll(ServerRequestInterface $request, ResponseInterface $response, array $args): ResponseInterface
     {
+        if(!$request->getHeader('admin_user_id')){
+            $response->getBody()->write(json_encode(["status"=>"error", "message"=>'O cabeçalho "admin_user_id" é obrigatório.']));
+            return $response->withStatus(400);
+        }
+        
         $adminUserId = $request->getHeader('admin_user_id')[0];
 
 
@@ -37,6 +42,11 @@ class ProductController
 
     public function lastUpdate(ServerRequestInterface $request, ResponseInterface $response, array $args): ResponseInterface
     {
+        if(!$request->getHeader('admin_user_id')){
+            $response->getBody()->write(json_encode(["status"=>"error", "message"=>'O cabeçalho "admin_user_id" é obrigatório.']));
+            return $response->withStatus(400);
+        }
+
         $stm = $this->service->getOne($args['id']);
         $product = Product::hydrateByFetch($stm->fetch());
 
@@ -96,6 +106,11 @@ class ProductController
     public function getOne(ServerRequestInterface $request, ResponseInterface $response, array $args): ResponseInterface
     {
         $stm = $this->service->getOne($args['id']);
+
+        if(!$stm->fetch()){
+            $response->getBody()->write(json_encode(["status"=>"error", "message"=>"Produto não encontrado!"]));
+            return $response->withStatus(404);
+        }
         $product = Product::hydrateByFetch($stm->fetch());
 
         $adminUserId = $request->getHeader('admin_user_id')[0];
@@ -117,35 +132,58 @@ class ProductController
 
     public function insertOne(ServerRequestInterface $request, ResponseInterface $response, array $args): ResponseInterface
     {
+
+        if(!$request->getHeader('admin_user_id')){
+            $response->getBody()->write(json_encode(["status"=>"error", "message"=>'O cabeçalho "admin_user_id" é obrigatório.']));
+            return $response->withStatus(400);
+        }
+
         $body = $request->getParsedBody();
         $adminUserId = $request->getHeader('admin_user_id')[0];
 
         if ($this->service->insertOne($body, $adminUserId)) {
+
+            $response->getBody()->write(json_encode(["status"=>"success", "message"=>"Produto cadastrado com sucesso!"]));
             return $response->withStatus(200);
         } else {
+            $response->getBody()->write(json_encode(["status"=>"error", "message"=>"Erro ao cadastrar produto"]));
             return $response->withStatus(404);
         }
     }
 
     public function updateOne(ServerRequestInterface $request, ResponseInterface $response, array $args): ResponseInterface
     {
+        if(!$request->getHeader('admin_user_id')){
+            $response->getBody()->write(json_encode(["status"=>"error", "message"=>'O cabeçalho "admin_user_id" é obrigatório.']));
+            return $response->withStatus(400);
+        }
+
         $body = $request->getParsedBody();
         $adminUserId = $request->getHeader('admin_user_id')[0];
 
         if ($this->service->updateOne($args['id'], $body, $adminUserId)) {
+            $response->getBody()->write(json_encode(["status"=>"success", "message"=>"Produto alterado com sucesso!"]));
             return $response->withStatus(200);
         } else {
+            $response->getBody()->write(json_encode(["status"=>"error", "message"=>"Erro ao alterar produto"]));
             return $response->withStatus(404);
         }
     }
 
     public function deleteOne(ServerRequestInterface $request, ResponseInterface $response, array $args): ResponseInterface
     {
+        if(!$request->getHeader('admin_user_id')){
+            $response->getBody()->write(json_encode(["status"=>"error", "message"=>'O cabeçalho "admin_user_id" é obrigatório.']));
+            return $response->withStatus(400);
+        }
+
         $adminUserId = $request->getHeader('admin_user_id')[0];
 
         if ($this->service->deleteOne($args['id'], $adminUserId)) {
+            $response->getBody()->write(json_encode(["status"=>"success", "message"=>"Produto removido com sucesso!"]));
             return $response->withStatus(200);
         } else {
+            $response->getBody()->write(json_encode(["status"=>"error", "message"=>"Erro ao remover produto"]));
             return $response->withStatus(404);
         }
     }
